+++ src/main/java/jscover/server/ConfigurationForServer.java
@@ -358,6 +358,7 @@ public class ConfigurationForServer extends ConfigurationCommon {
     public static final String DOC_ROOT_PREFIX = "--document-root=";
     public static final String PORT_PREFIX = "--port=";
     public static final String PROXY_PREFIX = "--proxy";
+    public static final String SAVE_JSON_ONLY = "--save-json-only";
     public static final String REPORT_DIR_PREFIX = "--report-dir=";
     public static final String INCLUDE_UNLOADED_JS_PREFIX = "--include-unloaded-js";
     public static final String URI_TO_FILE_MATCHER_PREFIX = "--uri-to-file-matcher=";
@@ -366,6 +367,7 @@ public class ConfigurationForServer extends ConfigurationCommon {
     private File documentRoot = new File(System.getProperty("user.dir"));
     private Integer port = 8080;
     private boolean proxy;
+    private boolean saveJSONOnly;
     private File reportDir = new File(System.getProperty("user.dir"));
     private boolean includeUnloadedJS;
     private Pattern uriToFileMatcher;
@@ -398,6 +400,10 @@ public class ConfigurationForServer extends ConfigurationCommon {
 
     public File getReportDir() {
         return reportDir;
+    }
+
+    public boolean isSaveJSONOnly() {
+        return saveJSONOnly;
     }
 
     public String getHelpText() {
@@ -440,6 +446,8 @@ public class ConfigurationForServer extends ConfigurationCommon {
             port = Integer.valueOf(arg.substring(PORT_PREFIX.length()));
         } else if (arg.equals(PROXY_PREFIX)) {
             proxy = true;
+        } else if (arg.equals(SAVE_JSON_ONLY)) {
+            saveJSONOnly = true;
         } else if (arg.equals(INCLUDE_UNLOADED_JS_PREFIX)) {
             includeUnloadedJS = true;
         } else if (arg.startsWith(REPORT_DIR_PREFIX)) {